log: rename striper file to view manager
