+++ src/libzlog/view_manager.h
